# scmOverrides: allow use of variables