Override scms having a common used set of overrides controlled by variables.

This reduces the number of overrides needed when you entered branching hell for your project... :see_no_evil: :hear_no_evil:

I'm not sure if it's a good idea to do the substitution only with the environment from `default.yaml` / command line but I have the feeling that substitution when constructing the `CheckoutStep` using the `fullEnv` is too late. It gives the recipes the possibility to change to override again which might be a pitfall.  :confused: