@jkloetzke 2018-06-30
I have thought about it for some time now and I think we should make the substitution late when creating the SCM object. That is in `ScmOverride.mangle()`. Otherwise, when doing the substitution just with the initial environment, the behavior will be inconsistent. Suppose you have the following recipe, where `BRANCH` is set by some other recipe (an not in the default environment):

```yaml
environment:
    FALLBACK_BRANCH: "master"

checkoutSCM:
    scm: git
    url: ...
    branch: "${BRANCH}"
```

What should be the result when applying the following override?

```yaml
scmOverrides:
    -
        match:
            branch: "${BRANCH}"
        set:
            branch: "${FALLBACK_BRANCH:-${BRANCH}}"
```

With the current approach you could neither match the branch (because the substitiution of the the attributes is done with the final environment) nor could you replace it with something from the final environment.

The drawback will be that the package calculation will slow down. But an inconsistent behavior would be worse IMO.