@rhubert 2018-07-09
Performance impact:

```
$ rm .bob-tree.* .bob-packages* 
$ time bob ls
```

### without Overrides

#### bob-master
```
real	1m25,753s
user	1m24,936s
sys	0m0,708s
```

#### this branch
```
real	1m26,179s
user	1m25,204s
sys	0m0,860s
```

### with Overrides (87 overrides active)

#### bob-master
```
real	1m27,200s
user	1m26,232s
sys	0m0,840s
```

#### this branch
```
real	1m31,959s
user	1m31,000s
sys	0m0,828s
```