@jkloetzke 2018-07-09
Looks good. The performance impact is there but IMHO in an acceptable range.

Is that something that needs to go into Bob 0.14? I'm a bit reluctant to merge new features during stabilization. OTOH it should not cause any harm...