@jkloetzke 2018-07-10
Ok, then we go for 0.14. Would you add some words in the respective chapter in the documentation and update the release notes? Then I'll merge it...